Added pre_comment_user_ip filter
At line 1095. This applies the filter (normally to anonymize the commenters IP address) when generating the hash to compare later. The hash is later compared with a newly generated one which uses the IP stored in the DB - which will have been (possibly) modified by the filter (as called by WP itself). If we don't do the filter when first generating the hash, then there will be a mismatch later and the user will not be given the edit possibility.